requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-taggit==6.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
packaging==25.0
pillow==11.2.1
python-telegram-bot==20.7
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.4.0
